+++ packages/gatsby-theme-basic/src/components/modularity-modules/posts-modules/ListPostsModule.js
@@ -1,4 +1,6 @@
+import { css } from "@emotion/react";
 import clsx from "clsx";
+import { kebabCase } from "lodash";
 import PropTypes from "prop-types";
 import React from "react";
 
@@ -39,8 +41,16 @@ export default function ListPostsModule({
       title={title}
       {...restProps}
       className={clsx(styles.component, theme, className)}
+      css={css({
+        "--list-rule-color": theme
+          ? `var(--brand-color-${kebabCase(theme)})`
+          : null,
+        "--module-wrapper-title-rule-color": theme
+          ? `var(--brand-color-${kebabCase(theme)})`
+          : null,
+      })}
     >
-      <RuledList className={clsx(styles.list)} ruleTop ruleBottom>
+      <RuledList className={clsx(styles.list)} ruleTop={!title} ruleBottom>
         {normalizedItems.map((item, index) => {
           return (
             <ListPostsModuleItem