+++ packages/gatsby-theme-basic/src/components/modularity-modules/posts-modules/index-layouts/index.js
@@ -1,2 +1,0 @@
-export { default as GridIndexPostsModule } from "./GridIndexPostsModule";
-export { default as MixedIndexPostsModule } from "./MixedIndexPostsModule";