+++ packages/gatsby-theme-basic/src/components/modularity-modules/posts-modules/index.js
@@ -1,5 +1,6 @@
 export { default as DefaultPostsModule } from "./DefaultPostsModule";
 export { default as ExpandableListPostsModule } from "./ExpandableListPostsModule";
-export { default as IndexPostsModule } from "./IndexPostsModule";
+export { default as HorizontalPostsModule } from "./HorizontalPostsModule";
+export { default as ItemsPostsModule } from "./ItemsPostsModule";
 export { default as ListPostsModule } from "./ListPostsModule";
 export { default as NewsPostsModule } from "./NewsPostsModule";