+++ packages/gatsby-theme-basic/src/components/modularity-modules/posts-modules/items/index.js
@@ -1,0 +1,2 @@
+export { default as PostsModuleBlocksItem } from "./PostsModuleBlocksItem";
+export { default as PostsModuleIndexItem } from "./PostsModuleIndexItem";