+++ packages/gatsby-theme-basic/src/hooks/useArchives.js
@@ -1,0 +1,11 @@
+import { useTranslation } from "react-i18next";
+
+export default function useArchives() {
+  const { t } = useTranslation();
+  return {
+    post: {
+      url: "/nyheter",
+      label: t("allPostsLabel"),
+    },
+  };
+}